Testing with model agent

+++ src/main/java/com/pitaev/leetcode/dp/optimization/BurstBalloons.java
@@ -10,7 +10,7 @@ package com.pitaev.leetcode.dp.optimization;
 public class BurstBalloons {
     public int maxCoins(int[] nums) {
         //update nums array to add 1's
-        int[] full = new int[nums.length + 2];
+        double[] full = new int[nums.length + 2];
         full[0] = 1;
         int i;
         for (i = 1; i <= nums.length; i++) {
